#55 - Support client generation with @Client and @Client.Import
Tidy imports

+++ http-api/src/test/java/io/avaje/http/api/PathTypeConversionTest.java
@@ -42,7 +42,7 @@ public class PathTypeConversionTest {
 
   @Test
   public void asLong() {
-    assertEquals(42L, PathTypeConversion.asInt("42"));
+    assertEquals(42L, PathTypeConversion.asLong("42"));
   }
 
   @Test
